Moved MenuByInput forward
