+++ layers/WPSchema/packages/menus/config/Overrides/schema-services.yaml
@@ -1,0 +1,10 @@
+services:
+    _defaults:
+        public: true
+        autowire: true
+
+    PoPSchema\Menus\TypeResolvers\InputObjectType\MenuByInputObjectTypeResolver:
+        class: PoPWPSchema\Menus\Overrides\TypeResolvers\InputObjectType\MenuByInputObjectTypeResolver
+
+    PoPSchema\Menus\FieldResolvers\ObjectType\RootObjectTypeFieldResolver:
+        class: PoPWPSchema\Menus\Overrides\FieldResolvers\ObjectType\RootObjectTypeFieldResolver
