+++ layers/WPSchema/packages/menus/config/schema-services.yaml
@@ -5,6 +5,3 @@ services:
 
     PoPWPSchema\Menus\FieldResolvers\:
         resource: '../src/FieldResolvers/*'
-
-    PoPWPSchema\Menus\TypeResolvers\:
-        resource: '../src/TypeResolvers/*'
