+++ layers/WPSchema/packages/menus/src/Component.php
@@ -35,5 +35,6 @@ class Component extends AbstractComponent
         array $skipSchemaComponentClasses = []
     ): void {
         self::initSchemaServices(dirname(__DIR__), $skipSchema);
+        self::initSchemaServices(dirname(__DIR__), $skipSchema, '/Overrides');
     }
 }
