--- layers/WPSchema/packages/menus/src/FieldResolvers/ObjectType/RootObjectTypeFieldResolver.php
@@ -1,149 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PoPWPSchema\Menus\FieldResolvers\ObjectType;
-
-use PoP\ComponentModel\FieldResolvers\ObjectType\AbstractQueryableObjectTypeFieldResolver;
-use PoP\ComponentModel\Schema\SchemaTypeModifiers;
-use PoP\ComponentModel\TypeResolvers\ConcreteTypeResolverInterface;
-use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
-use PoP\Engine\TypeResolvers\ObjectType\RootObjectTypeResolver;
-use PoP\Engine\TypeResolvers\ScalarType\StringScalarTypeResolver;
-use PoPSchema\Menus\TypeResolvers\ObjectType\MenuObjectTypeResolver;
-use PoPWPSchema\Menus\TypeResolvers\InputObjectType\MenuByInputObjectTypeResolver;
-
-/**
- * Override implementation of field "menu" to provide a MenuByInput with several options
- */
-class RootObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolver
-{
-    private ?MenuObjectTypeResolver $menuObjectTypeResolver = null;
-    private ?StringScalarTypeResolver $stringScalarTypeResolver = null;
-    private ?MenuByInputObjectTypeResolver $menuByInputObjectTypeResolver = null;
-
-    final public function setMenuObjectTypeResolver(MenuObjectTypeResolver $menuObjectTypeResolver): void
-    {
-        $this->menuObjectTypeResolver = $menuObjectTypeResolver;
-    }
-    final protected function getMenuObjectTypeResolver(): MenuObjectTypeResolver
-    {
-        return $this->menuObjectTypeResolver ??= $this->instanceManager->getInstance(MenuObjectTypeResolver::class);
-    }
-    final public function setStringScalarTypeResolver(StringScalarTypeResolver $stringScalarTypeResolver): void
-    {
-        $this->stringScalarTypeResolver = $stringScalarTypeResolver;
-    }
-    final protected function getStringScalarTypeResolver(): StringScalarTypeResolver
-    {
-        return $this->stringScalarTypeResolver ??= $this->instanceManager->getInstance(StringScalarTypeResolver::class);
-    }
-    final public function setMenuByInputObjectTypeResolver(MenuByInputObjectTypeResolver $menuByInputObjectTypeResolver): void
-    {
-        $this->menuByInputObjectTypeResolver = $menuByInputObjectTypeResolver;
-    }
-    final protected function getMenuByInputObjectTypeResolver(): MenuByInputObjectTypeResolver
-    {
-        return $this->menuByInputObjectTypeResolver ??= $this->instanceManager->getInstance(MenuByInputObjectTypeResolver::class);
-    }
-
-    public function getObjectTypeResolverClassesToAttachTo(): array
-    {
-        return [
-            RootObjectTypeResolver::class,
-        ];
-    }
-
-    public function getFieldNamesToResolve(): array
-    {
-        return [
-            'menu',
-        ];
-    }
-
-    /**
-     * Execute before "menu" in Menus
-     */
-    public function getPriorityToAttachToClasses(): int
-    {
-        return 100;
-    }
-
-    public function getFieldDescription(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName): ?string
-    {
-        return match ($fieldName) {
-            'menu' => $this->getTranslationAPI()->__('Get a menu by some property', 'menus'),
-            default => parent::getFieldDescription($objectTypeResolver, $fieldName),
-        };
-    }
-
-    public function getFieldArgNameTypeResolvers(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName): array
-    {
-        $fieldArgNameTypeResolvers = parent::getFieldArgNameTypeResolvers($objectTypeResolver, $fieldName);
-        return match ($fieldName) {
-            'menu' => array_merge(
-                $fieldArgNameTypeResolvers,
-                [
-                    'by' => $this->getMenuByInputObjectTypeResolver(),
-                ]
-            ),
-            default => $fieldArgNameTypeResolvers,
-        };
-    }
-
-    public function getFieldArgTypeModifiers(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName, string $fieldArgName): int
-    {
-        return match ([$fieldName => $fieldArgName]) {
-            ['menu' => 'by'] => SchemaTypeModifiers::MANDATORY,
-            default => parent::getFieldArgTypeModifiers($objectTypeResolver, $fieldName, $fieldArgName),
-        };
-    }
-
-    /**
-     * @param array<string, mixed> $fieldArgs
-     * @param array<string, mixed>|null $variables
-     * @param array<string, mixed>|null $expressions
-     * @param array<string, mixed> $options
-     */
-    public function resolveValue(
-        ObjectTypeResolverInterface $objectTypeResolver,
-        object $object,
-        string $fieldName,
-        array $fieldArgs,
-        ?array $variables = null,
-        ?array $expressions = null,
-        array $options = []
-    ): mixed {
-        switch ($fieldName) {
-            case 'menu':
-                $menuParam = null;
-                $by = $fieldArgs['by'];
-                if (isset($by->id)) {
-                    $menuParam = $by->id;
-                } elseif (isset($by->slug)) {
-                    $menuParam = $by->slug;
-                } elseif (isset($by->location)) {
-                    $locations = \get_nav_menu_locations();
-                    $menuParam = $locations[$by->location] ?? null;
-                }
-                if ($menuParam === null) {
-                    return null;
-                }
-                $menu = wp_get_nav_menu_object($menuParam);
-                if ($menu === false) {
-                    return null;
-                }
-                return $menu->term_id;
-        }
-
-        return parent::resolveValue($objectTypeResolver, $object, $fieldName, $fieldArgs, $variables, $expressions, $options);
-    }
-
-    public function getFieldTypeResolver(ObjectTypeResolverInterface $objectTypeResolver, string $fieldName): ConcreteTypeResolverInterface
-    {
-        return match ($fieldName) {
-            'menu' => $this->getMenuObjectTypeResolver(),
-            default => parent::getFieldTypeResolver($objectTypeResolver, $fieldName),
-        };
-    }
-}
